requirements.txt
streamlit
requests
opencv-python
decord
Pillow
numpy